refactor: eliminate massive duplication across Circuit classes
- Extract common initialization into Circuit::Base module
- Create shared Circuit::StateCallbacks for duplicated callback methods
- Refactor Circuit, AsyncCircuit, and CoordinatedCircuit to use inheritance
- Add JRuby test helper for Rails-dependent test skipping
