+++ lib/breaker_machines.rb
@@ -14,6 +14,7 @@ loader.ignore("#{__dir__}/breaker_machines/types.rb")
 loader.ignore("#{__dir__}/breaker_machines/console.rb")
 loader.ignore("#{__dir__}/breaker_machines/async_support.rb")
 loader.ignore("#{__dir__}/breaker_machines/hedged_async_support.rb")
+loader.ignore("#{__dir__}/breaker_machines/circuit/async_state_management.rb")
 loader.setup
 
 # BreakerMachines provides a thread-safe implementation of the Circuit Breaker pattern
