+++ lib/breaker_machines/circuit.rb
@@ -1,14 +1,10 @@
 # frozen_string_literal: true
 
-require 'concurrent-ruby'
+require_relative 'circuit/base'
 
 module BreakerMachines
   class Circuit
+    include Circuit::Base
     include StateManagement
-    include Configuration
-    include Execution
-    include Circuit::HedgedExecution
-    include Introspection
-    include Callbacks
   end
 end
