+++ lib/breaker_machines/circuit/configuration.rb
@@ -11,30 +11,6 @@ module BreakerMachines
         attr_reader :name, :config, :opened_at
       end
 
-      def initialize(name, options = {})
-        @name = name
-        @config = default_config.merge(options)
-        # Always use a storage backend for proper sliding window implementation
-        # Use global default storage if not specified
-        @storage = @config[:storage] || create_default_storage
-        @metrics = @config[:metrics]
-        @opened_at = Concurrent::AtomicReference.new(nil)
-        @half_open_attempts = Concurrent::AtomicFixnum.new(0)
-        @half_open_successes = Concurrent::AtomicFixnum.new(0)
-        @mutex = Concurrent::ReentrantReadWriteLock.new
-        @last_failure_at = Concurrent::AtomicReference.new(nil)
-        @last_error = Concurrent::AtomicReference.new(nil)
-
-        # Initialize semaphore for bulkheading if max_concurrent is set
-        @semaphore = (Concurrent::Semaphore.new(@config[:max_concurrent]) if @config[:max_concurrent])
-
-        super() # Initialize state machine
-        restore_status_from_storage if @storage
-
-        # Register with global registry unless auto_register is disabled
-        BreakerMachines::Registry.instance.register(self) unless @config[:auto_register] == false
-      end
-
       private
 
       def default_config
