+++ lib/breaker_machines/circuit/introspection.rb
@@ -6,6 +6,7 @@ module BreakerMachines
     # and generating human-readable summaries of circuit status.
     module Introspection
       extend ActiveSupport::Concern
+
       # State check methods are automatically generated by state_machines:
       # - open? (returns true when status == :open)
       # - closed? (returns true when status == :closed)
