+++ test/test_helper.rb
@@ -7,7 +7,11 @@ require 'minitest/autorun'
 require 'active_support/test_case'
 
 # Require async for fiber tests
-require 'async'
+begin
+  require 'async'
+rescue LoadError
+  # Async not available (e.g., on JRuby)
+end
 
 # Suppress notifications during tests unless explicitly testing them
 BreakerMachines.config.log_events = false
@@ -19,33 +23,45 @@ BreakerMachines.config.fiber_safe = false
 BreakerMachines.logger = Logger.new(nil)
 
 # Load Rails environment
-ENV['RAILS_ENV'] ||= 'test'
-require_relative 'dummy/config/environment'
-require 'rails/test_help'
-require 'state_machines/integrations/active_record'
+begin
+  ENV['RAILS_ENV'] ||= 'test'
+  require_relative 'dummy/config/environment'
+  require 'rails/test_help'
+  require 'state_machines/integrations/active_record'
 
-# Load the schema
-ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: ':memory:')
-ActiveRecord::Schema.verbose = false
-load File.expand_path('dummy/db/schema.rb', __dir__)
+  # Load the schema
+  ActiveRecord::Base.establish_connection(adapter: 'sqlite3', database: ':memory:')
+  ActiveRecord::Schema.verbose = false
+  load File.expand_path('dummy/db/schema.rb', __dir__)
+rescue LoadError
+  # Rails not available (e.g., on JRuby)
+  require_relative 'jruby_test_helper'
+end
 
-# Global teardown to ensure circuit registry is cleared between ALL tests
-module ActiveSupport
-  class TestCase
-    # Disable parallel tests globally to prevent DRb connection pool corruption
-    # with FallbackChain and other storage timeout behaviors
-    parallelize(workers: 1)
+# Skip Rails-dependent tests helper
+module RailsTestSkipper
+  def skip_rails_dependent_test
+    # This is a no-op when Rails is available
   end
 end
 
-module Minitest
-  class Test
-    def teardown
-      super
-      # Clear the circuit registry after every test to prevent state leakage
-      BreakerMachines.registry.clear
-      # Clear Rails cache to prevent storage state leakage
-      Rails.cache.clear
+# Global teardown to ensure circuit registry is cleared between ALL tests
+if defined?(ActiveSupport)
+  module ActiveSupport
+    class TestCase
+      include RailsTestSkipper
+
+      # Disable parallel tests globally to prevent DRb connection pool corruption
+      # with FallbackChain and other storage timeout behaviors
+      parallelize(workers: 1)
+
+      def teardown
+        super
+        # Clear the circuit registry after every test to prevent state leakage
+        BreakerMachines.registry.clear
+        # Clear Rails cache to prevent storage state leakage
+        Rails.cache.clear if defined?(Rails)
+      end
     end
   end
 end
